initial test fixes
